+++ packages/vendure-plugin-invoices/src/ui/invoices.module.ts
@@ -2,7 +2,6 @@ import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { SharedModule } from '@vendure/admin-ui/core';
 import { InvoicesComponent } from './invoices.component';
-import { InvoiceDetailViewComponent } from './invoices-detail-view/invoices-detail-view.component';
 
 @NgModule({
   imports: [
@@ -17,6 +16,6 @@ import { InvoiceDetailViewComponent } from './invoices-detail-view/invoices-deta
     ]),
   ],
   providers: [],
-  declarations: [InvoicesComponent, InvoiceDetailViewComponent],
+  declarations: [InvoicesComponent],
 })
 export class InvoicesModule {}