+++ packages/vendure-plugin-address-lookup/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 2.1.0 (2026-02-05)
+
+- Upgraded to Vendure 3.5.3
+
+# 2.0.1 (2026-02-04)
+
+- Log addditional error details for BAG API
+
 # 2.0.0 (2025-12-24)
 
 - Replaced Postcode.tech API with BAG api, because Postcode.tech is no longer available