+++ packages/vendure-plugin-modify-customer-orders/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.7.0 (2026-02-05)
+
+- Upgraded to Vendure 3.5.3
+
 # 1.6.2 (2025-11-13)
 
 - Documentation update